Move remove subscription logic to subscription handler
